rename the file name tab2 into Clickcounter.tsx and also its code
